+++ test/mocks/mock_local_storage.dart
@@ -1,5 +1,0 @@
-// test/mocks/mock_local_storage.dart
-import 'package:flutter_corekit/src/storage/secure_storage.dart';
-import 'package:mocktail/mocktail.dart';
-
-class MockLocalStorage extends Mock implements SecureStorage {}